Rename "testing_utils" to "cpu_utils"
Because it's not only used for testing.
